requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
distro
httpx<0.24
langserve